finish moving functions to view model
